Replace stdlib usage with lib
