+++ compiler-plugin-test/src/test/resources/testng.xml
@@ -21,7 +21,7 @@
 <suite name="BallerinaJsondataCompilerPluginTests">
     <test name="UnitTests">
         <classes>
-            <class name="io.ballerina.stdlib.data.jsondata.compiler.CompilerPluginTest"/>
+            <class name="io.ballerina.lib.data.jsondata.compiler.CompilerPluginTest"/>
         </classes>
     </test>
 </suite>
