--- compiler-plugin/src/main/java/io/ballerina/stdlib/data/jsondata/compiler/JsondataDiagnosticCodes.java
@@ -1,57 +1,0 @@
-/*
- * Copyright (c) 2024, WSO2 LLC. (https://www.wso2.com).
- *
- * WSO2 LLC. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package io.ballerina.stdlib.data.jsondata.compiler;
-
-import io.ballerina.tools.diagnostics.DiagnosticSeverity;
-
-import static io.ballerina.tools.diagnostics.DiagnosticSeverity.ERROR;
-
-/**
- * Diagnostic codes for Jsondata's compiler plugin.
- *
- * @since 0.1.0
- */
-public enum JsondataDiagnosticCodes {
-
-    UNSUPPORTED_UNION_TYPE("JSON_ERROR_201",
-            "unsupported union type: union type does not support multiple complex types", ERROR),
-    DUPLICATE_FIELD("JSON_ERROR_202", "invalid field: duplicate field found", ERROR);
-
-    private final String code;
-    private final String message;
-    private final DiagnosticSeverity severity;
-
-    JsondataDiagnosticCodes(String code, String message, DiagnosticSeverity severity) {
-        this.code = code;
-        this.message = message;
-        this.severity = severity;
-    }
-
-    public String getCode() {
-        return code;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-
-    public DiagnosticSeverity getSeverity() {
-        return severity;
-    }
-}
