--- native/src/main/java/io/ballerina/stdlib/data/jsondata/io/DataReaderThreadPool.java
@@ -1,55 +1,0 @@
-/*
- *  Copyright (c) 2024, WSO2 LLC. (https://www.wso2.com).
- *
- *  WSO2 LLC. licenses this file to you under the Apache License,
- *  Version 2.0 (the "License"); you may not use this file except
- *  in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing,
- *  software distributed under the License is distributed on an
- *  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- *  KIND, either express or implied.  See the License for the
- *  specific language governing permissions and limitations
- *  under the License.
- */
-package io.ballerina.stdlib.data.jsondata.io;
-
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.SynchronousQueue;
-import java.util.concurrent.ThreadFactory;
-import java.util.concurrent.ThreadPoolExecutor;
-import java.util.concurrent.TimeUnit;
-
-/**
- * Thread pool for data reader.
- *
- * @since 0.1.0
- */
-public class DataReaderThreadPool {
-
-    // TODO : Make this configurable, in Ballerina Library.
-    private static final int CORE_POOL_SIZE = 0;
-    private static final int MAX_POOL_SIZE = 50;
-    private static final long KEEP_ALIVE_TIME = 60L;
-    private static final String THREAD_NAME = "bal-data-jsondata-thread";
-    public static final ExecutorService EXECUTOR_SERVICE = new ThreadPoolExecutor(CORE_POOL_SIZE,
-            MAX_POOL_SIZE, KEEP_ALIVE_TIME, TimeUnit.SECONDS, new SynchronousQueue<>(), new DataThreadFactory());
-
-    /**
-     * Thread factory for data reader.
-     *
-     * @since 0.1.0
-     */
-    static class DataThreadFactory implements ThreadFactory {
-
-        @Override
-        public Thread newThread(Runnable runnable) {
-            Thread ballerinaData = new Thread(runnable);
-            ballerinaData.setName(THREAD_NAME);
-            return ballerinaData;
-        }
-    }
-}
